Refactor components into their own directories Components now have their own directories to themselves main.tsx is updated to use the correct import
