+++ src/components/InfoCard/InfoCard.tsx
