+++ src/components/InfoCard/index.ts
@@ -1,0 +1,1 @@
+export { default as InfoCard } from "./InfoCard";
