+++ src/components/Scramble/index.ts
@@ -1,0 +1,1 @@
+export { default as Scramble } from "./Scramble";
