+++ src/components/Timer/Timer.tsx
