+++ src/components/Timer/index.ts
@@ -1,0 +1,1 @@
+export { default as Timer } from "./Timer";
